refactor pecPorts to its own module
