+++ src/javaharness/java/arcs/webimpl/PortableJsonJsImpl.java
@@ -10,7 +10,6 @@ import java.util.Set;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import jsinterop.base.Any;
-import java.lang.ClassCastException;
 
 public class PortableJsonJsImpl implements PortableJson {
     private Any jsonObj;
